Add createNewColumn function to streamline custom field creation and validation

+++ admin/modules/system/custom_field.php
@@ -51,6 +51,22 @@ if (!$can_read && $_SESSION['uid'] != 1) {
     die('<div class="errorBox">'.__('You don\'t have enough privileges to access this area!').'</div>');
 }
 
+function createNewColumn($tableName, $columnName, $columnType) {
+    Schema::table($tableName, function(Blueprint $table) use ($columnName, $columnType) {
+        if ($columnType == 'text') {
+            $table->string($columnName, 255)->nullable()->add();
+        } elseif ($columnType == 'longtext') {
+            $table->text($columnName)->nullable()->add();
+        } elseif ($columnType == 'numeric') {
+            $table->integer($columnName, 11)->nullable()->add();
+        } elseif ($columnType == 'dropdown' || $columnType == 'checklist' || $columnType == 'choice') {
+            $table->string($columnName, 255)->nullable()->add();
+        } elseif ($columnType == 'date') {
+            $table->date($columnName)->nullable()->add();
+        }
+    });
+}
+
 // set table options
 $table_options = array();
 $table_options[] = array('biblio', __('Bibliography'));
@@ -95,16 +111,28 @@ if (isset($_POST['saveData']) AND $can_read AND $can_write) {
             /* UPDATE RECORD MODE */
             // filter update record ID
             $updateRecordID = $dbs->escape_string(trim($_POST['updateRecordID']));
+
             //get last field table
             $_q = $dbs->query("SELECT primary_table,dbfield,label FROM mst_custom_field WHERE field_id=".$updateRecordID);
             if($_q->num_rows){
                 $_d = $_q->fetch_row();
                 if($_d[0]!=$data['primary_table']){
-                    // drop column
-                    $schemaParams = [$_d[0].'_custom', $_d[1]];
-                    if (Schema::hasColumn(...$schemaParams)) Schema::dropColumn(...$schemaParams);
+                    $count_q = $dbs->query("SELECT COUNT(*) FROM ".$_d[0]."_custom WHERE ".$_d[1]." IS NOT NULL");
+                    $count_d = $count_q->fetch_row();
+                    $has_data = ($count_d[0]) < 1;
+                    if(!$has_data){
+                        utility::jsToastr(__('Custom Field'), __('You can not change primary table if the field already has data, ' . $count_d[0] . ' row(s) found.'), 'error');
+                        exit();
+                    } else {
+                        $schemaParams = [$_d[0].'_custom', $_d[1]];
+                        if (Schema::hasColumn(...$schemaParams)) Schema::dropColumn(...$schemaParams);
+                        // create new column
+                        $data['dbfield'] = 'cf_'.substr(md5(microtime()),rand(0,26),5);
+                        createNewColumn($data['primary_table'].'_custom', $data['dbfield'], $data['type']);
+                    }
                 }
             }
+
             // update the data
             $update = $sql_op->update('mst_custom_field', $data, 'field_id=\''.$updateRecordID.'\'');
             if ($update) {
@@ -122,24 +150,12 @@ if (isset($_POST['saveData']) AND $can_read AND $can_write) {
             
             $tableName = $data['primary_table'].'_custom';
             $schemaParams = [$tableName, $data['dbfield']];
-            
-            Schema::table($tableName, function(Blueprint $table) use ($data) {
-                if($data['type'] == 'text'){
-                    $table->string($data['dbfield'], 255)->nullable()->add();
-                }else if($data['type'] == 'longtext'){
-                    $table->text($data['dbfield'])->nullable()->add();
-                }else if($data['type'] == 'numeric'){
-                    $table->integer($data['dbfield'], 11)->nullable()->add();
-                }else if($data['type'] == 'dropdown'){
-                    $table->string($data['dbfield'], 255)->nullable()->add();
-                }else if($data['type'] == 'checklist'){
-                    $table->string($data['dbfield'], 255)->nullable()->add();
-                }else if($data['type'] == 'choice'){
-                    $table->string($data['dbfield'], 255)->nullable()->add();
-                }else if($data['type'] == 'date'){
-                    $table->date($data['dbfield'])->nullable()->add();
-                }
-            });
+
+            // check if the column already exists
+            if (!Schema::hasColumn(...$schemaParams)) {
+                // create new column
+                createNewColumn($tableName, $data['dbfield'], $data['type']);
+            }
 
             $insert = $sql_op->insert('mst_custom_field', $data);
             if ($insert) {
